dfu: add optional timeout
When the `dfu` command is called from the U-Boot environment, it now
accepts an optional fourth parameter that specifies a timeout (in
seconds). If a DFU connection is not made within that time the dfu
command exits (as it would if ctrl+c was pressed). If the timeout is
left empty the dfu command behaves as it does now.

This is useful for allowing U-Boot to check to see if anything wants to
upload new firmware before continuing to boot.

This option is not compatible with the dfu "list" command. If "list" is
specified that command takes priority.

This patch is based on commit 5e966ccc3c65c18c9783741fa04e0c45e021780c
from
https://github.com/01org/edison-u-boot/commit/5e966ccc3c65c18c9783741fa04e0c45e021780c
It updates that patch to the newer changes of u-boot which split the DFU
functionality into multiple functions.

Signed-off-by: Sebastien Colleur <sebastienx.colleur@intel.com>
Signed-off-by: Brad Campbell <bradjc5@gmail.com>