Makefile.lib: Always rebuild DSDT
The dsdt.asl is usually combined out of several files that are included
in the main one. Whenever we change the content of any of such files,
build system is not able to recognize them. Hence the easiest way is to
force DSDT rebuild each time we run make.

Note it mimics the behaviour for Device Tree compilation.

Signed-off-by: Andy Shevchenko <andriy.shevchenko@linux.intel.com>